+++ client-reactjs/src/components/application/dashboard/Orbit/Orbit.jsx
@@ -8,21 +8,25 @@ import Filters from './Filters';
 import MetricBoxes from '../common/charts/MetricBoxes';
 import '../common/css/index.css';
 import ExportMenu from '../ExportMenu/ExportMenu';
-// import { withTheme } from 'styled-components';
+import { authHeader, handleError } from '../../../common/AuthHeader.js';
+import { message } from 'antd';
 
 function Orbit() {
+  //all states needed to manage data
   const [builds, setBuilds] = useState([]);
   const [workUnits, setWorkUnits] = useState([]);
-  const [loadingData, setLoadingData] = useState(true);
+  const [filteredWorkUnits, setFilteredWorkUnits] = useState(null);
+  const [filteredBuilds, setFilteredBuilds] = useState(null);
+
+  //states for charts and titles
   const [titleMetrics, setTitleMetrics] = useState([]);
   const [metrics, setMetrics] = useState([]);
   const [stackBarData, setStackBarData] = useState([]);
   const [donutData, setDonutData] = useState([]);
-  const [severity, setSeverity] = useState(['0', '1', '2', '3']);
   const [groupDataBy, setGroupDataBy] = useState('day');
+
+  //states for filters
   const [dashboardFilters, setDashboardFilters] = useState({});
-  const [filteredWorkUnits, setFilteredWorkUnits] = useState([]);
-  const [filteredBuilds, setFilteredBuilds] = useState([]);
   const [filterValues, setFilterValues] = useState({
     initialStatus: [],
     initialStatusOptions: [],
@@ -51,26 +55,295 @@ function Orbit() {
     ],
   });
 
+  //states for spinners
+  const [loading, setLoading] = useState(false);
+  const [loadingData, setLoadingData] = useState(true);
+
+  //grab the application ID from redux
   const {
     application: { applicationId },
   } = useSelector((item) => item.applicationReducer);
 
+  // useEffect(() => {
+  //   if (filteredWorkUnits !== null) {
+  //     const groupedData = filteredWorkUnits.map((workUnit) => {
+  //       const weekStart = moment(workUnit.lastRun).startOf('week').format('MM/DD/YY');
+  //       const weekEnd = moment(workUnit.lastRun).endOf('week').format('MM/DD/YY');
+  //       const updatedItem = { ...workUnit };
+  //       updatedItem.metaData.lastRun = `${weekStart} - ${weekEnd}`;
+  //       return updatedItem;
+  //     });
+  //     setStackBarData(groupedData);
+
+  //     console.log(groupedData);
+  //   }
+  // }, [groupDataBy]);
+
+  // Step 1 - first we are loading the initial builds and workunits into state
   useEffect(() => {
-    const groupedData = builds.map((build) => {
-      const weekStart = moment(build.Date).startOf('week').format('MM/DD/YY');
-      const weekEnd = moment(build.Date).endOf('week').format('MM/DD/YY');
-      const updatedItem = { ...build };
-      updatedItem.Date = `${weekStart} - ${weekEnd}`;
-      return updatedItem;
+    getbuilds();
+  }, [applicationId]);
+
+  //Get list of all builds, workunits, and set initial values and filters
+  const getbuilds = async () => {
+    try {
+      setLoading(true);
+      const payload = {
+        method: 'GET',
+        header: authHeader(),
+      };
+
+      if (applicationId === undefined) return;
+      var startTime2 = performance.now();
+      const response = await fetch(`/api/orbit/allMonitoring/${applicationId}`, payload);
+      if (!response.ok) handleError(response);
+      const data = await response.json();
+      var endTime2 = performance.now();
+      console.log(`getting builds took ${endTime2 - startTime2} milliseconds`);
+
+      var startTime3 = performance.now();
+      //get work unit information and put it in builds information
+      const response2 = await fetch(`/api/orbit/getWorkUnits/${applicationId}`, payload);
+      if (!response2.ok) handleError(response2);
+      const data2 = await response2.json();
+      var endTime3 = performance.now();
+      console.log(`getting workunits took ${endTime3 - startTime3} milliseconds`);
+
+      let builds2 = [];
+      var startTime = performance.now();
+      await Promise.all(
+        //add data2 workunits to matching builds
+        (builds2 = data.map((build) => {
+          const wu = data2.filter((workUnit) => workUnit.name === build.build);
+          return { ...build, workUnits: wu };
+        }))
+      );
+
+      //combine and set work units
+      let totalWuList = data2;
+
+      totalWuList.sort((a, b) => {
+        return new Date(b.metaData.lastRun) - new Date(a.metaData.lastRun);
+      });
+
+      //move initial status, final status, and version to top level of object
+      totalWuList.forEach((workUnit) => {
+        workUnit.key = workUnit.id;
+        workUnit.initialStatus = workUnit.metaData.initialStatus;
+        workUnit.finalStatus = workUnit.metaData.finalStatus;
+        workUnit.version = workUnit.metaData.version;
+        workUnit.status = workUnit.metaData.status;
+        workUnit.lastRun = workUnit.metaData.lastRun;
+      });
+
+      //get all the unique values for the filters and dropdowns
+
+      //get unique values from workunit list
+      const uniqueInitialStatus = [...new Set(totalWuList.map((item) => item.initialStatus.toUpperCase()))];
+      const uniqueFinalStatus = [...new Set(totalWuList.map((item) => item.finalStatus.toUpperCase()))];
+      const uniqueVersion = [...new Set(totalWuList.map((item) => item.version))];
+      const uniqueWorkUnits = [...new Set(totalWuList.map((item) => item.wuid))];
+
+      //get unique values from builds list
+      const uniqueSeverity = [...new Set(builds2.map((item) => item.severityCode))];
+      const uniqueBuilds = [...new Set(builds2.map((item) => item.build))];
+      const uniqueProducts = [...new Set(builds2.map((item) => item.product))];
+      const uniqueBusinessUnits = [...new Set(builds2.map((item) => item.businessUnit))];
+
+      //create options for dropdowns
+      const uniqueInitialStatusOptions = [];
+      uniqueInitialStatus.forEach((item) => uniqueInitialStatusOptions.push({ label: item, value: item }));
+      const uniqueFinalStatusOptions = [];
+      uniqueFinalStatus.forEach((item) => uniqueFinalStatusOptions.push({ label: item, value: item }));
+      const uniqueVersionOptions = [];
+      uniqueVersion.forEach((item) => uniqueVersionOptions.push({ label: item, value: item }));
+      const uniqueSeverityOptions = [];
+      uniqueSeverity.forEach((item) => uniqueSeverityOptions.push({ label: item, value: item }));
+      const uniqueBuildsOptions = [];
+      uniqueBuilds.forEach((item) => uniqueBuildsOptions.push({ label: item, value: item }));
+      const uniqueProductsOptions = [];
+      uniqueProducts.forEach((item) => uniqueProductsOptions.push({ label: item, value: item }));
+      const uniqueBusinessUnitsOptions = [];
+      uniqueBusinessUnits.forEach((item) => uniqueBusinessUnitsOptions.push({ label: item, value: item }));
+      const uniqueWorkUnitsOptions = [];
+      uniqueWorkUnits.forEach((item) => uniqueWorkUnitsOptions.push({ label: item, value: item }));
+
+      //check URL params for filters
+      const params = new URLSearchParams(location.search);
+      const filters = {};
+      if (params.get('initialStatus')) {
+        filters.initialStatus = params.get('initialStatus')?.split(',');
+      }
+      if (params.get('finalStatus')) {
+        filters.finalStatus = params.get('finalStatus')?.split(',');
+      }
+      if (params.get('dateRange')) {
+        const dateString = params.get('dateRange');
+        const dates = dateString.split(',');
+        const range = [moment(dates[0]), moment(dates[1])];
+        filters.dateRange = range;
+      }
+      if (params.get('groupDataBy')) {
+        filters.groupDataBy = params.get('groupDataBy');
+        setGroupDataBy(params.get('groupDataBy'));
+        // setGroupDataBy('day');
+      }
+
+      setWorkUnits(totalWuList);
+
+      setBuilds(builds2);
+
+      //set them in state for the initial load
+      setFilterValues((filterValues) => ({
+        ...filterValues,
+        initialStatus: filters.initialStatus?.length ? filters.initialStatus : uniqueInitialStatus,
+        initialStatusOptions: uniqueInitialStatusOptions,
+        finalStatus: filters.finalStatus?.length ? filters.finalStatus : uniqueFinalStatus,
+        finalStatusOptions: uniqueFinalStatusOptions,
+        version: uniqueVersion,
+        versionOptions: uniqueVersionOptions,
+        severity: uniqueSeverity,
+        severityOptions: uniqueSeverityOptions,
+        builds: uniqueBuilds,
+        buildsOptions: uniqueBuildsOptions,
+        products: uniqueProducts,
+        productsOptions: uniqueProductsOptions,
+        businessUnits: uniqueBusinessUnits,
+        businessUnitsOptions: uniqueBusinessUnitsOptions,
+        wuid: uniqueWorkUnits,
+        wuidOptions: uniqueWorkUnitsOptions,
+        dateRange: filters.dateRange?.length ? filters.dateRange : [moment().subtract(15, 'days'), moment()],
+        groupDataBy: filters.groupDataBy ? filters.groupDataBy : 'day',
+      }));
+
+      var endTime = performance.now();
+
+      console.log(`setting initial facets took ${endTime - startTime} milliseconds`);
+    } catch (error) {
+      message.error('Failed to fetch builds' + error);
+      console.log(error);
+    }
+  };
+
+  //Step 2 - now we are setting the intial filters
+  useEffect(() => {
+    if (!Object.keys(dashboardFilters).length && filterValues.initialStatus.length > 0) {
+      setDashboardFilters(filterValues);
+      return;
+    }
+  }, [filterValues]);
+
+  // Step 3 - now we need to apply the initial filters to the initial data
+  useEffect(() => {
+    //if filtered builds and workunits are empty and the initial data and filters are set, filter data
+    if (
+      filteredBuilds === null &&
+      filteredWorkUnits === null &&
+      builds.length > 0 &&
+      workUnits.length > 0 &&
+      Object.keys(dashboardFilters).length > 0
+    ) {
+      filterData();
+    }
+  }, [builds, workUnits, dashboardFilters]);
+
+  const filterData = () => {
+    let startTime = performance.now();
+    let filteredBuildsList = [];
+
+    if (builds.length === 0 || workUnits.length === 0) return;
+    //apply filters to build list
+    builds.forEach((build) => {
+      if (
+        dashboardFilters.severity.includes(build.severityCode) &&
+        dashboardFilters.builds.includes(build.build) &&
+        dashboardFilters.products.includes(build.product) &&
+        dashboardFilters.businessUnits.includes(build.businessUnit)
+      ) {
+        filteredBuildsList.push(build);
+      } else {
+        return;
+      }
     });
 
-    setStackBarData(groupedData);
-  }, [groupDataBy]);
+    setFilteredBuilds(filteredBuildsList);
 
-  // When build changes run
+    //get a list of just the names to filter work units
+    let filteredBuildNameList = filteredBuildsList.map((build) => build.build);
+
+    //apply filters to work units once the build list is filtered
+    let filtered = workUnits.filter((workUnit) => {
+      let wuDate = moment(workUnit.lastRun);
+
+      if (
+        wuDate > moment(dashboardFilters.dateRange[0]) &&
+        wuDate < moment(dashboardFilters.dateRange[1]) &&
+        dashboardFilters.initialStatus?.includes(workUnit.initialStatus.toUpperCase()) &&
+        dashboardFilters.finalStatus?.includes(workUnit.finalStatus.toUpperCase()) &&
+        dashboardFilters.version?.includes(workUnit.version) &&
+        filteredBuildNameList.includes(workUnit.name)
+      ) {
+        return true;
+      } else {
+        return false;
+      }
+    });
+
+    setFilteredWorkUnits(filtered);
+    let endTime = performance.now();
+    console.log(`filtering data took ${endTime - startTime} milliseconds`);
+  };
+
+  //Step 4 - now that the filtered builds and workunits are set, we can get the counts and chart data
   useEffect(() => {
+    if (filteredBuilds !== null && filteredWorkUnits !== null && filteredBuilds[0].count === undefined) {
+      getCounts();
+      setChartData();
+    }
+  }, [filteredBuilds, filteredWorkUnits]);
+
+  const getCounts = () => {
+    let startTime = performance.now();
+    let builds2 = [];
+    let wus = [];
+
+    filteredBuilds.forEach((build) => {
+      let count = 0;
+
+      if (build.workUnits?.length > 0) {
+        build.workUnits.forEach((workUnit) => {
+          let wuDate = moment(workUnit.lastRun);
+
+          if (
+            wuDate > moment(dashboardFilters.dateRange[0]) &&
+            wuDate < moment(dashboardFilters?.dateRange[1]) &&
+            dashboardFilters.initialStatus?.includes(workUnit.initialStatus.toUpperCase()) &&
+            dashboardFilters.finalStatus?.includes(workUnit.finalStatus.toUpperCase()) &&
+            dashboardFilters.version?.includes(workUnit.version) &&
+            dashboardFilters.builds?.includes(workUnit.name)
+          ) {
+            //check if workunit already exists in wus list to avoid duplicates
+            let wuExists = wus.find((item) => item.id === workUnit.id);
+            if (!wuExists) {
+              wus.push(workUnit);
+              count++;
+            }
+          }
+        });
+      }
+
+      builds2.push({ ...build, count: count });
+    });
+
+    setFilteredBuilds(builds2);
+    setLoading(false);
+    let endTime = performance.now();
+    console.log(`getting counts took ${endTime - startTime} milliseconds`);
+  };
+
+  const setChartData = () => {
+    let startTime = performance.now();
     if (filteredWorkUnits) {
-      var startTime = performance.now();
       const newMetrics = []; // Pie data
       const newStackBarData = []; // Stack bar Data
       const newDonutData = []; // Donut data
@@ -100,8 +373,8 @@ function Orbit() {
       switch (groupDataBy) {
         case 'week':
           data = filteredWorkUnits.map((workUnit) => {
-            const weekStart = moment(workUnit.metaData.lastRun).startOf('week').format('MM/DD/YY');
-            const weekEnd = moment(workUnit.metaData.lastRun).endOf('week').format('MM/DD/YY');
+            const weekStart = moment(workUnit.lastRun).startOf('week').format('MM/DD/YY');
+            const weekEnd = moment(workUnit.lastRun).endOf('week').format('MM/DD/YY');
             const updatedItem = { ...workUnit };
             updatedItem.metaData.lastRun = `${weekStart} - ${weekEnd}`;
             return updatedItem;
@@ -111,9 +384,7 @@ function Orbit() {
         case 'month':
           data = filteredWorkUnits.map((workUnit) => {
             const updatedItem = { ...workUnit };
-            updatedItem.metaData.lastRun = moment(moment(workUnit.metaData.lastRun).utc(), 'MM/DD/YYYY').format(
-              'MMMM YYYY'
-            );
+            updatedItem.metaData.lastRun = moment(moment(workUnit.lastRun).utc(), 'MM/DD/YYYY').format('MMMM YYYY');
             return updatedItem;
           });
           break;
@@ -121,7 +392,7 @@ function Orbit() {
         case 'quarter':
           data = filteredWorkUnits.map((workUnit) => {
             const updatedItem = { ...workUnit };
-            const date = moment.utc(workUnit.metaData.lastRun);
+            const date = moment.utc(workUnit.lastRun);
             const year = date.year();
             const quarter = Math.ceil((date.month() + 1) / 3);
             updatedItem.metaData.lastRun = `${year} - Q${quarter}`;
@@ -132,7 +403,7 @@ function Orbit() {
         case 'year':
           data = filteredWorkUnits.map((workUnit) => {
             const updatedItem = { ...workUnit };
-            const date = moment.utc(workUnit.metaData.lastRun);
+            const date = moment.utc(workUnit.lastRun);
             const year = date.year();
             updatedItem.metaData.lastRun = year;
             return updatedItem;
@@ -196,11 +467,18 @@ function Orbit() {
       setMetrics(newMetrics);
       setStackBarData(newStackBarData);
       setDonutData(newDonutData);
-      var endTime = performance.now();
-
-      console.log(`setting visible chart data takes ${endTime - startTime} milliseconds`);
+      let endTime = performance.now();
+      console.log(`setting chart data took ${endTime - startTime} milliseconds`);
     }
-  }, [builds, filteredBuilds, filteredWorkUnits, groupDataBy]);
+  };
+
+  //Step 5 (repeatable) - now we need to re-render the data when it filters change
+  useEffect(() => {
+    //if dashboard filters are set and filtered builds and workunits
+    if (Object.keys(dashboardFilters).length && filteredBuilds !== null && filteredWorkUnits !== null) {
+      filterData();
+    }
+  }, [dashboardFilters]);
 
   return (
     <div>
@@ -217,8 +495,6 @@ function Orbit() {
                 isOrbit={true}
                 dashboardFilters={dashboardFilters}
                 setDashboardFilters={setDashboardFilters}
-                severity={severity}
-                setSeverity={setSeverity}
                 filterValues={filterValues}
               />
             </Space>
@@ -238,21 +514,7 @@ function Orbit() {
             />
           </div>
           <div style={{ width: '100%', display: 'flex', justifyContent: 'space-between' }}>
-            <OrbitTable
-              applicationId={applicationId}
-              dashboardFilters={dashboardFilters}
-              builds={builds}
-              setBuilds={setBuilds}
-              workUnits={workUnits}
-              setWorkUnits={setWorkUnits}
-              filteredWorkUnits={filteredWorkUnits}
-              setFilteredWorkUnits={setFilteredWorkUnits}
-              filteredBuilds={filteredBuilds}
-              setFilteredBuilds={setFilteredBuilds}
-              filterValues={filterValues}
-              setFilterValues={setFilterValues}
-              setGroupDataBy={setGroupDataBy}
-            />
+            <OrbitTable filteredWorkUnits={filteredWorkUnits} filteredBuilds={filteredBuilds} loading={loading} />
 
             {builds.length > 0 ? (
               <div className="builds__charts">
